requirements.txt
spoon-ai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
web3>=6.11.0
eth-account>=0.10.0
pydantic>=2.5.0
aiohttp>=3.9.0
jinja2>=3.1.2
python-multipart>=0.0.6
hashlib
typing-extensions>=4.9.0
